Whack a mole game
This is a whack a mole game based on tinyOS and python game. In this
project, four telosb mote are used as moles, one pop router is used as
controller. The game is controlled by a python game, which is used to
receive the input data and display the result. The demo of this project
can be checked in
https://www.youtube.com/watch?v=AAMDzTGoMbw&spfreload=10